requirements.txt
numpy
scikit-learn
torch
pytest
matplotlib
pyyaml
torchinfo
pykan